+++ lib/nit_riverpod_notifications.dart
@@ -4,5 +4,4 @@ export 'src/nit_notifications/nit_notification.dart';
 export 'src/nit_notifications/nit_notification_types_enum.dart';
 
 export 'src/ref_extension.dart';
-export 'src/widget_ref_extension.dart';
 export 'src/nit_notification_listener_widget.dart';