+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+Basic logic released:
+- can use NitNotificationListenerWidget and ref.notifyUser
+- can use default NitNotification class and respective notificationPresenter